initial work on listen to fb friends on newhive dialog
this commit involves major changes to facebook authorization code.  we
are now able to share a session started by the client (javascript sdk)
with the server.  This is a much nicer flow, and I want to apply it
elsewhere in the fb integration code.  some serious cleanup and
refactoring still needs to be done