[Delivers #75481292] Categories backend
UI for adding/editing/removing catories is available when logged in as root.
To add a collection to a category, hit the aqua + button on the profile when viewing a collection.
To update categories, navigate to root's profile, and click the aqua # button.  Reordering the categories and the collections within a category is the same UI as it is for tags and collections.